virtualbox/requirements.txt
Click==7.0
Jinja2==2.10.3
PyYAML==5.1.2
